% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{add_qq_confidence_band}
\alias{add_qq_confidence_band}
\title{Add Confidence Band to QQ Plot}
\usage{
add_qq_confidence_band(n, conf_level = 0.95)
}
\description{
Add Confidence Band to QQ Plot
}
\keyword{internal}
