% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_utilities.R
\name{add_ridge}
\alias{add_ridge}
\title{Add Ridge to Matrix Diagonal}
\usage{
add_ridge(K, ridge = 0.001)
}
\arguments{
\item{K}{Square matrix}

\item{ridge}{Ridge value to add to diagonal}
}
\value{
Matrix with ridge added to diagonal
}
\description{
Adds a small value to the diagonal of a matrix for numerical stability
}
\examples{
\dontrun{
K <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
K_ridge <- add_ridge(K, ridge = 0.01)
}
}
