% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrices.R
\name{align_matrices}
\alias{align_matrices}
\title{Align Matrices to Common Individuals}
\usage{
align_matrices(K_list, phenotypes, id_col = "GID")
}
\arguments{
\item{K_list}{List of relationship matrices}

\item{phenotypes}{Phenotype data frame}

\item{id_col}{Name of ID column in phenotypes (default "GID")}
}
\value{
List with aligned matrices and phenotypes
}
\description{
Ensures all matrices and data have the same individuals in the same order
}
\examples{
\dontrun{
K <- calc_relationship_matrices(gdata, matrices = "A")
aligned <- align_matrices(K, gdata@phenotypes)
}
}
