% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{apply_genomic_control}
\alias{apply_genomic_control}
\title{Apply Genomic Control Correction}
\usage{
apply_genomic_control(gwas_result)
}
\arguments{
\item{gwas_result}{GWASResult object}
}
\value{
GWASResult object with corrected p-values
}
\description{
Apply Genomic Control Correction
}
