% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_qtl_aware_gblup.R
\name{build_design_matrices_qtl}
\alias{build_design_matrices_qtl}
\title{Build Design Matrices for QTL-Aware Model}
\usage{
build_design_matrices_qtl(pheno, K_matrices, effects)
}
\description{
Build Design Matrices for QTL-Aware Model
}
\keyword{internal}
