% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_matrices.R
\name{build_E_matrix}
\alias{build_E_matrix}
\title{Build Epistatic Relationship Matrix}
\usage{
build_E_matrix(
  M,
  type = c("A#A", "A#D", "D#D"),
  A = NULL,
  D = NULL,
  min.MAF = 0.01,
  impute.method = c("mean"),
  use.cpp = TRUE
)
}
\arguments{
\item{M}{Marker matrix (individuals x markers) coded as -1, 0, 1}

\item{type}{Type of epistasis ("A#A", "A#D", or "D#D")}

\item{A}{Optional pre-computed A matrix}

\item{D}{Optional pre-computed D matrix}

\item{min.MAF}{Minimum minor allele frequency for filtering markers}

\item{impute.method}{Method for imputing missing values ("mean")}

\item{use.cpp}{Use C++ implementation for speed (default TRUE)}
}
\value{
Epistatic relationship matrix
}
\description{
Constructs epistatic relationship matrices (A#A, A#D, or D#D) using
Hadamard (element-wise) product with fast C++ implementation.
}
