% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_design_matrices.R
\name{build_Z_matrix_cpp}
\alias{build_Z_matrix_cpp}
\title{Build Z Matrix Using C++ (Fast Version)}
\usage{
build_Z_matrix_cpp(factor_vector, levels)
}
\arguments{
\item{factor_vector}{Vector of factor values for each observation}

\item{levels}{Character vector of all levels (in desired order)}
}
\value{
Sparse design matrix
}
\description{
Creates a sparse incidence matrix using C++ for maximum speed.
}
\keyword{internal}
