% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrices.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Calculate Genomic Prediction Accuracy}
\usage{
calc_accuracy(observed, predicted, method = c("pearson", "spearman"))
}
\arguments{
\item{observed}{Vector of observed values}

\item{predicted}{Vector of predicted values}

\item{method}{Correlation method ("pearson", "spearman")}
}
\value{
Correlation coefficient
}
\description{
Computes correlation between observed and predicted values
}
\examples{
\dontrun{
observed <- c(1, 2, 3, 4, 5)
predicted <- c(1.1, 2.2, 2.9, 4.1, 4.8)
calc_accuracy(observed, predicted)
}
}
