% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_D_matrix_cpp}
\alias{calc_D_matrix_cpp}
\title{Fast Dominance Relationship Matrix (Vitezica 2013)}
\usage{
calc_D_matrix_cpp(G)
}
\arguments{
\item{G}{Genotype matrix (individuals x markers) coded as -1, 0, 1}
}
\value{
Dominance relationship matrix (n x n)
}
\description{
Calculates the dominance relationship matrix using Vitezica et al. (2013) method.
}
