% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_epistatic_matrix_cpp}
\alias{calc_epistatic_matrix_cpp}
\title{Fast Epistatic Relationship Matrix (Hadamard Product)}
\usage{
calc_epistatic_matrix_cpp(A, B)
}
\arguments{
\item{A}{First relationship matrix}

\item{B}{Second relationship matrix}
}
\value{
Epistatic relationship matrix (Hadamard product)
}
\description{
Calculates epistatic relationship matrix as element-wise product of two matrices.
}
