% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_matrices.R
\name{calc_inbreeding}
\alias{calc_inbreeding}
\title{Calculate Inbreeding Coefficients from Relationship Matrix}
\usage{
calc_inbreeding(K)
}
\arguments{
\item{K}{Relationship matrix}
}
\value{
Vector of inbreeding coefficients
}
\description{
Calculate Inbreeding Coefficients from Relationship Matrix
}
