% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrices.R
\name{calc_mae}
\alias{calc_mae}
\title{Calculate Mean Absolute Error}
\usage{
calc_mae(observed, predicted)
}
\arguments{
\item{observed}{Vector of observed values}

\item{predicted}{Vector of predicted values}
}
\value{
MAE value
}
\description{
Calculate Mean Absolute Error
}
