% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrices.R
\name{calc_mse}
\alias{calc_mse}
\title{Calculate Mean Squared Error}
\usage{
calc_mse(observed, predicted)
}
\arguments{
\item{observed}{Vector of observed values}

\item{predicted}{Vector of predicted values}
}
\value{
MSE value
}
\description{
Calculate Mean Squared Error
}
\examples{
\dontrun{
observed <- c(1, 2, 3, 4, 5)
predicted <- c(1.1, 2.2, 2.9, 4.1, 4.8)
calc_mse(observed, predicted)
}
}
