% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrices.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Calculate Root Mean Squared Error}
\usage{
calc_rmse(observed, predicted)
}
\arguments{
\item{observed}{Vector of observed values}

\item{predicted}{Vector of predicted values}
}
\value{
RMSE value
}
\description{
Calculate Root Mean Squared Error
}
