% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\alias{calculate_accuracy,GBLUPModel-method}
\title{Calculate Predictive Ability (Correlation)}
\usage{
calculate_accuracy(object)

\S4method{calculate_accuracy}{GBLUPModel}(object)
}
\arguments{
\item{object}{A GBLUPModel object}
}
\value{
Numeric correlation coefficient between observed and predicted values
}
\description{
Calculates correlation between observed phenotypes and predicted GEBVs (predictive ability).
}
