% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{Calculate Genomic Inflation Factor (Lambda)}
\usage{
calculate_lambda(p_values)
}
\description{
Calculate Genomic Inflation Factor (Lambda)
}
\keyword{internal}
