% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_openmp_support}
\alias{check_openmp_support}
\title{Check OpenMP Support}
\usage{
check_openmp_support()
}
\value{
Logical indicating if OpenMP is available
}
\description{
Check OpenMP Support
}
\keyword{internal}
