% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_cross_validation.R
\name{compile_predictions}
\alias{compile_predictions}
\title{Compile Predictions from All Folds}
\usage{
compile_predictions(all_results)
}
\arguments{
\item{all_results}{List of results from all replications}
}
\value{
Data frame with all predictions
}
\description{
Compile Predictions from All Folds
}
\keyword{internal}
