% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_design_matrices.R
\name{compute_ZKZt}
\alias{compute_ZKZt}
\title{Compute ZKZ' Efficiently}
\usage{
compute_ZKZt(Z, K)
}
\arguments{
\item{Z}{Incidence matrix (can be sparse)}

\item{K}{Relationship matrix}
}
\value{
ZKZ' matrix
}
\description{
Computes Z * K * Z' using either dense or sparse operations
depending on input types.
}
\keyword{internal}
