% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_ZKZt_cpp}
\alias{compute_ZKZt_cpp}
\title{Efficient computation of ZKZ'}
\usage{
compute_ZKZt_cpp(Z, K)
}
\arguments{
\item{Z}{Incidence matrix (can be sparse)}

\item{K}{Relationship matrix (dense)}
}
\value{
ZKZ' matrix
}
\description{
Computes Z * K * Z' efficiently for variance component models.
}
