% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_cross_validation.R
\name{create_env_folds}
\alias{create_env_folds}
\title{Create Environment Folds for CV2}
\usage{
create_env_folds(pheno)
}
\arguments{
\item{pheno}{Phenotype data frame}
}
\value{
List of fold indices
}
\description{
CV2 scheme: Leave-one-environment-out cross-validation.
Tests ability to predict in completely new environments.
}
\keyword{internal}
