% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_design_matrices.R
\name{create_model_matrix}
\alias{create_model_matrix}
\title{Create Model Matrix from Formula}
\usage{
create_model_matrix(formula, data)
}
\arguments{
\item{formula}{Formula for fixed effects}

\item{data}{Data frame}
}
\value{
Design matrix
}
\description{
Creates fixed effects design matrix from a formula and data frame.
}
