% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_Z_sparse_cpp}
\alias{create_Z_sparse_cpp}
\title{Sparse incidence matrix creation}
\usage{
create_Z_sparse_cpp(factor_idx, n_levels)
}
\arguments{
\item{factor_idx}{Integer vector of factor levels (1-indexed)}

\item{n_levels}{Number of unique levels}
}
\value{
Sparse matrix (dgCMatrix format)
}
\description{
Creates a sparse incidence (Z) matrix mapping observations to levels.
}
