% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{CVResult-class}
\alias{CVResult-class}
\title{CVResult Class}
\description{
S4 class to store cross-validation results
}
\section{Slots}{

\describe{
\item{\code{predictions}}{Data frame with all predictions across folds}

\item{\code{metrics}}{Data frame with metrics by fold}

\item{\code{fold_results}}{List of detailed results by fold}

\item{\code{scheme}}{CV scheme used}

\item{\code{n_folds}}{Number of folds}

\item{\code{n_reps}}{Number of replications}

\item{\code{effects}}{Effects included in model}
}}

