% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_bed_cpp}
\alias{decode_bed_cpp}
\title{Fast PLINK BED File Decoding}
\usage{
decode_bed_cpp(bed_data, n_ind, n_snp)
}
\arguments{
\item{bed_data}{Raw vector containing BED file bytes (after 3-byte header)}

\item{n_ind}{Number of individuals}

\item{n_snp}{Number of SNPs}
}
\value{
Integer matrix (n_ind x n_snp) with genotypes coded as 0, 1, 2, or NA
}
\description{
Decodes PLINK binary BED format genotype data to integer matrix.
Uses standard PLINK BED encoding:
\itemize{
\item 00 (binary) = 0 (homozygous reference)
\item 01 (binary) = NA (missing)
\item 10 (binary) = 1 (heterozygous)
\item 11 (binary) = 2 (homozygous alternate)
}
}
