% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_design_matrices.R
\name{design_matrices}
\alias{design_matrices}
\title{Design Matrix Construction}
\description{
Functions to build design matrices for mixed models with optional
sparse matrix support for memory efficiency.
}
\keyword{internal}
