% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emma_reml_loglik_cpp}
\alias{emma_reml_loglik_cpp}
\title{Fast REML log-likelihood evaluation (EMMA-style)}
\usage{
emma_reml_loglik_cpp(eta, xi, d, lambda)
}
\arguments{
\item{eta}{Transformed response (U'y where U is eigenvector matrix)}

\item{xi}{Transformed fixed effects (U'X)}

\item{d}{Eigenvalues of K matrix}

\item{lambda}{Ratio of residual to genetic variance (sigma2_e / sigma2_g)}
}
\value{
REML log-likelihood value
}
\description{
Computes the REML log-likelihood given transformed data and lambda.
Uses pre-computed eigendecomposition for O(n) complexity per evaluation.
}
