% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{estimate_variance_components}
\alias{estimate_variance_components}
\title{Estimate Variance Components using REML}
\usage{
estimate_variance_components(
  y,
  X,
  K,
  method = "REML",
  max_iter = 100,
  tol = 1e-06,
  verbose = FALSE
)
}
\description{
Estimate Variance Components using REML
}
\keyword{internal}
