% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_qtl_aware_gblup.R
\name{extract_gebvs_qtl}
\alias{extract_gebvs_qtl}
\title{Extract GEBVs from QTL-Aware Model}
\usage{
extract_gebvs_qtl(
  fit,
  pheno,
  design,
  X,
  Ka,
  has_qtl,
  qtl_effects,
  gdata = NULL
)
}
\description{
Extract GEBVs from QTL-Aware Model
}
\keyword{internal}
