% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_matrices.R
\name{filter_markers_maf}
\alias{filter_markers_maf}
\title{Filter Markers by Minor Allele Frequency}
\usage{
filter_markers_maf(M, min.MAF = 0.01, verbose = TRUE)
}
\arguments{
\item{M}{Marker matrix}

\item{min.MAF}{Minimum minor allele frequency}
}
\value{
Filtered marker matrix
}
\description{
Filter Markers by Minor Allele Frequency
}
\keyword{internal}
