% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_henderson_mme_cpp}
\alias{fit_henderson_mme_cpp}
\title{Henderson's MME Solver with EM-REML (C++ Version)}
\usage{
fit_henderson_mme_cpp(
  y,
  X,
  Z,
  K,
  sigma2_g_init = -1,
  sigma2_e_init = -1,
  max_iter = 50L,
  tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{y}{Response vector (n x 1)}

\item{X}{Fixed effects design matrix (n x p)}

\item{Z}{Random effects design matrix (n x q)}

\item{K}{Relationship matrix (q x q)}

\item{sigma2_g_init}{Initial genetic variance}

\item{sigma2_e_init}{Initial residual variance}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
List with model components including variance components
}
\description{
Complete Henderson's MME solver with EM-REML variance component estimation.
Supports single random effect for now (can be extended to multiple).
}
