% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_mme_solver.R
\name{fit_mme_ai_reml}
\alias{fit_mme_ai_reml}
\title{Full AI-REML Algorithm for Multiple Random Effects}
\usage{
fit_mme_ai_reml(y, X, Z_list, K_list, nIters, tolParConvLL, tolParInv, verbose)
}
\description{
Full AI-REML Algorithm for Multiple Random Effects
}
\keyword{internal}
