% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{GBLUPData-class}
\alias{GBLUPData-class}
\title{GBLUPData Class}
\description{
S4 class to store genotype and phenotype data for GBLUP analysis
}
\section{Slots}{

\describe{
\item{\code{genotypes}}{Matrix of genotype data (individuals x markers), coded as -1, 0, 1}

\item{\code{phenotypes}}{Data frame with columns: GID (genotype ID), ENV (environment), trait (phenotypic value)}

\item{\code{maf}}{Numeric vector of minor allele frequencies}

\item{\code{metadata}}{List of additional metadata}
}}

