% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_cross_validation.R
\name{get_cv_metrics}
\alias{get_cv_metrics}
\title{Extract Cross-Validation Metrics}
\usage{
get_cv_metrics(cv_result, by_fold = FALSE)
}
\arguments{
\item{cv_result}{CVResult object}

\item{by_fold}{Logical, return metrics by fold}
}
\value{
Data frame of metrics
}
\description{
Extract Cross-Validation Metrics
}
\examples{
\dontrun{
cv_res <- cv_gblup(gdata, K, effects = "A", scheme = "CV1")
get_cv_metrics(cv_res)
get_cv_metrics(cv_res, by_fold = TRUE)
}
}
