% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrices.R
\name{get_genotypes}
\alias{get_genotypes}
\title{Extract Genotype Matrix from GBLUPData}
\usage{
get_genotypes(gdata, individuals = NULL)
}
\arguments{
\item{gdata}{GBLUPData object}

\item{individuals}{Optional vector of individual IDs to extract}
}
\value{
Genotype matrix
}
\description{
Extract Genotype Matrix from GBLUPData
}
\examples{
\dontrun{
gdata <- simulate_genotypes(n_ind = 100, n_snp = 500)
M <- get_genotypes(gdata)
M_subset <- get_genotypes(gdata, individuals = c("ID1", "ID2", "ID3"))
}
}
