% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrices.R
\name{get_phenotypes}
\alias{get_phenotypes}
\title{Extract Phenotype Data from GBLUPData}
\usage{
get_phenotypes(gdata, trait = NULL, environment = NULL)
}
\arguments{
\item{gdata}{GBLUPData object}

\item{trait}{Optional trait name to extract}

\item{environment}{Optional environment to filter}
}
\value{
Data frame of phenotypes
}
\description{
Extract Phenotype Data from GBLUPData
}
\examples{
\dontrun{
gdata <- simulate_genotypes(n_ind = 100, n_snp = 500)
gdata <- simulate_phenotypes(gdata, n_env = 3, h2 = 0.6)
pheno <- get_phenotypes(gdata)
pheno_e1 <- get_phenotypes(gdata, environment = "E1")
}
}
