% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{get_top_markers}
\alias{get_top_markers}
\title{Get Top Markers from GWAS}
\usage{
get_top_markers(gwas_result, n = 20, threshold = NULL)
}
\arguments{
\item{gwas_result}{GWASResult object}

\item{n}{Number of top markers to return}

\item{threshold}{Optional p-value threshold}
}
\value{
Data frame with top markers
}
\description{
Get Top Markers from GWAS
}
