% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_model_fitting.R
\name{get_varcomp}
\alias{get_varcomp}
\title{Get variance components from model}
\usage{
get_varcomp(model)
}
\arguments{
\item{model}{GBLUPModel object}
}
\value{
Data frame of variance components with columns Component and Variance
}
\description{
Get variance components from model
}
\examples{
\dontrun{
model <- fit_gblup(gdata, K, effects = "A")
vc <- get_varcomp(model)
print(vc)
}
}
