% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{gwas}
\alias{gwas}
\title{Genome-Wide Association Study (GWAS)}
\usage{
gwas(
  gdata,
  K_matrices = NULL,
  effects = "A",
  min.MAF = 0.05,
  P3D = TRUE,
  n.PC = 0,
  method = c("Wald", "LRT"),
  n.cores = 1,
  use.cpp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gdata}{GBLUPData object}

\item{K_matrices}{List of relationship matrices for population structure control}

\item{effects}{Character vector of random effects to include in the model}

\item{min.MAF}{Minimum minor allele frequency for filtering markers}

\item{P3D}{Logical, if TRUE variance components estimated once (default TRUE)}

\item{n.PC}{Number of principal components to include as fixed effects}

\item{method}{Method for p-value calculation ("Wald" or "LRT")}

\item{n.cores}{Number of cores for parallel processing (default 1)}

\item{use.cpp}{Use C++ implementation for speed (default TRUE)}

\item{verbose}{Logical, print progress messages}
}
\value{
GWASResult object
}
\description{
Performs GWAS using EMMAX-style efficient mixed linear model with
spectral decomposition for fast marker testing.
}
\examples{
\dontrun{
gdata <- simulate_genotypes(n_ind = 200, n_snp = 1000)
gdata <- simulate_phenotypes(gdata, n_env = 1, h2 = 0.6, n_qtl = 10)
K <- calc_relationship_matrices(gdata, matrices = "A")
gwas_res <- gwas(gdata, K_matrices = K)
plot_manhattan(gwas_res)
plot_qq(gwas_res)
}
}
