% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{GWASResult-class}
\alias{GWASResult-class}
\title{GWASResult Class}
\description{
S4 class to store GWAS results
}
\section{Slots}{

\describe{
\item{\code{results}}{Data frame with marker test results}

\item{\code{lambda}}{Genomic inflation factor}

\item{\code{n_markers}}{Number of markers tested}

\item{\code{min_MAF}}{Minimum MAF used}

\item{\code{method}}{Test method used}
}}

