% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_matrices.R
\name{impute_markers}
\alias{impute_markers}
\title{Impute Missing Marker Values}
\usage{
impute_markers(M, method = c("mean"))
}
\arguments{
\item{M}{Marker matrix with possible missing values}

\item{method}{Imputation method ("mean")}
}
\value{
Imputed marker matrix
}
\description{
Imputes missing values in marker matrix using specified method.
}
\keyword{internal}
