% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_matrices.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{Check if Matrix is Positive Definite}
\usage{
is_positive_definite(K, tol = 1e-08, use.cpp = TRUE)
}
\arguments{
\item{K}{Square matrix}

\item{tol}{Tolerance for smallest eigenvalue}

\item{use.cpp}{Use C++ implementation (default TRUE)}
}
\value{
Logical, TRUE if positive definite
}
\description{
Check if Matrix is Positive Definite
}
