% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_positive_definite_cpp}
\alias{is_positive_definite_cpp}
\title{Check if matrix is positive definite}
\usage{
is_positive_definite_cpp(M, tol = 1e-08)
}
\arguments{
\item{M}{Square matrix}

\item{tol}{Tolerance for smallest eigenvalue}
}
\value{
TRUE if positive definite
}
\description{
Check if matrix is positive definite
}
