% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_pd_cpp}
\alias{make_pd_cpp}
\title{Make matrix positive definite by adding ridge}
\usage{
make_pd_cpp(M, tol = 1e-06)
}
\arguments{
\item{M}{Square matrix}

\item{tol}{Tolerance for smallest eigenvalue}
}
\value{
Positive definite matrix
}
\description{
Make matrix positive definite by adding ridge
}
