% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{plot_qq}
\alias{plot_qq}
\title{QQ Plot for GWAS Results}
\usage{
plot_qq(gwas_result, main = "QQ Plot", ...)
}
\arguments{
\item{gwas_result}{GWASResult object}

\item{main}{Plot title}

\item{...}{Additional arguments for plot}
}
\value{
Invisible NULL
}
\description{
QQ Plot for GWAS Results
}
