% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_visualization.R
\name{plot_relationship_matrix}
\alias{plot_relationship_matrix}
\title{Plot Relationship Matrix Heatmap}
\usage{
plot_relationship_matrix(
  K,
  cluster = TRUE,
  main = "Genomic Relationship Matrix",
  col = NULL,
  max_ind = 100,
  ...
)
}
\arguments{
\item{K}{Relationship matrix}

\item{cluster}{Cluster individuals}

\item{main}{Plot title}

\item{col}{Color palette}

\item{max_ind}{Maximum individuals to plot}

\item{...}{Additional arguments}
}
\value{
Invisible NULL
}
\description{
Plot Relationship Matrix Heatmap
}
