% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_visualization.R
\name{plot_selection_response}
\alias{plot_selection_response}
\title{Plot Selection Response}
\usage{
plot_selection_response(
  gebv,
  top_percent = 0.2,
  main = "Selection Response",
  ...
)
}
\arguments{
\item{gebv}{Data frame with GEBV column}

\item{top_percent}{Percentage of top individuals}

\item{main}{Plot title}

\item{...}{Additional arguments}
}
\value{
Invisible NULL
}
\description{
Plot Selection Response
}
