% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{predict,GBLUPModel-method}
\alias{predict,GBLUPModel-method}
\title{Predict method for GBLUPModel}
\usage{
\S4method{predict}{GBLUPModel}(object, ...)
}
\arguments{
\item{object}{A GBLUPModel object}

\item{...}{Additional arguments (not used)}
}
\value{
Data frame with predictions
}
\description{
Predict method for GBLUPModel
}
