% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_mme_solver.R
\name{predict_blup_new}
\alias{predict_blup_new}
\title{Predict BLUPs for New Individuals Using MME Framework}
\usage{
predict_blup_new(u_train, K, train_ids, new_ids)
}
\arguments{
\item{u_train}{BLUPs for training individuals}

\item{K}{Full relationship matrix (includes both train and new)}

\item{train_ids}{IDs of training individuals}

\item{new_ids}{IDs of new individuals to predict}
}
\value{
Vector of predicted BLUPs for new individuals
}
\description{
Computes proper BLUP predictions for individuals not in training set.
}
