% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{show-methods}
\alias{show-methods}
\alias{show,GBLUPData-method}
\alias{show,GBLUPModel-method}
\alias{show,CVResult-method}
\alias{show,GWASResult-method}
\title{Show Methods for CompGEBLUP Objects}
\usage{
\S4method{show}{GBLUPData}(object)

\S4method{show}{GBLUPModel}(object)

\S4method{show}{CVResult}(object)

\S4method{show}{GWASResult}(object)
}
\arguments{
\item{object}{An S4 object (GBLUPData, GBLUPModel, CVResult, or GWASResult)}
}
\description{
Display summary information for S4 classes
}
