% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_generation.R
\name{simulate_genotypes}
\alias{simulate_genotypes}
\title{Simulate Genotype Data}
\usage{
simulate_genotypes(n_ind, n_snp, maf_range = c(0.05, 0.5))
}
\arguments{
\item{n_ind}{Number of individuals}

\item{n_snp}{Number of SNPs}

\item{maf_range}{MAF range (default c(0.05, 0.5))}
}
\value{
GBLUPData object with simulated genotypes
}
\description{
Generates a simulated genotype matrix with specified dimensions and MAF range.
}
\examples{
\dontrun{
gdata <- simulate_genotypes(n_ind = 100, n_snp = 1000)
show(gdata)
}
}
