% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_generation.R
\name{simulate_phenotypes}
\alias{simulate_phenotypes}
\title{Simulate Phenotypes}
\usage{
simulate_phenotypes(
  gblup_data,
  n_env = 3,
  h2 = 0.6,
  n_qtl = 50,
  include_dominance = FALSE,
  include_epistasis = FALSE,
  include_gxe = TRUE,
  mu = 100,
  prop_dom = 0.2,
  prop_epi = 0.15,
  prop_gxe = 0.1,
  prop_env = 0.2
)
}
\arguments{
\item{gblup_data}{GBLUPData object}

\item{n_env}{Number of environments}

\item{h2}{Heritability (0-1)}

\item{n_qtl}{Number of causal QTL}

\item{include_dominance}{Include dominance effects}

\item{include_epistasis}{Include epistatic effects (AA)}

\item{include_gxe}{Include GxE effects}

\item{mu}{Overall mean}

\item{prop_dom}{Proportion of genetic variance due to dominance (default 0.2)}

\item{prop_epi}{Proportion of genetic variance due to epistasis (default 0.15)}

\item{prop_gxe}{Proportion of genetic variance due to GxE (default 0.1)}

\item{prop_env}{Proportion of environmental variance due to environment main effect (default 0.2)}
}
\value{
GBLUPData object with simulated phenotypes
}
\description{
Simulates phenotypic values based on genotype data with specified genetic architecture.
}
\examples{
\dontrun{
gdata <- simulate_genotypes(n_ind = 100, n_snp = 500)
gdata <- simulate_phenotypes(gdata, n_env = 3, h2 = 0.6, n_qtl = 20)
head(gdata@phenotypes)
}
}
