% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_mme_cpp}
\alias{solve_mme_cpp}
\title{Solve mixed model equations efficiently}
\usage{
solve_mme_cpp(y, X, Z, K, sigma2_g, sigma2_e)
}
\arguments{
\item{y}{Response vector}

\item{X}{Fixed effects design matrix}

\item{Z}{Random effects design matrix}

\item{K}{Relationship matrix}

\item{sigma2_g}{Genetic variance}

\item{sigma2_e}{Residual variance}
}
\value{
List with fixed effects (b) and random effects (u)
}
\description{
Solves the Henderson's mixed model equations using efficient algorithms.
}
