% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_mme_solver.R
\name{solve_mme_with_varcomp}
\alias{solve_mme_with_varcomp}
\title{Solve MME with Known Variance Components}
\usage{
solve_mme_with_varcomp(y, X, Z, K, sigma2_g, sigma2_e, ridge = 1e-06)
}
\description{
Solve MME with Known Variance Components
}
\keyword{internal}
