% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_cross_validation.R
\name{summarize_cv_results}
\alias{summarize_cv_results}
\title{Summarize Cross-Validation Results}
\usage{
summarize_cv_results(all_results, scheme)
}
\arguments{
\item{all_results}{List of results from all replications}

\item{scheme}{CV scheme used}
}
\value{
List with summary statistics
}
\description{
Summarize Cross-Validation Results
}
\keyword{internal}
