% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{summary,CVResult-method}
\alias{summary,CVResult-method}
\title{Summary Method for CVResult}
\usage{
\S4method{summary}{CVResult}(object, ...)
}
\arguments{
\item{object}{CVResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Summary Method for CVResult
}
