% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{summary,GWASResult-method}
\alias{summary,GWASResult-method}
\title{Summary Method for GWASResult}
\usage{
\S4method{summary}{GWASResult}(object, ...)
}
\arguments{
\item{object}{GWASResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Summary Method for GWASResult
}
