% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_utilities.R
\name{summary_gblup_data}
\alias{summary_gblup_data}
\title{Summary Statistics for GBLUPData}
\usage{
summary_gblup_data(gblup_data)
}
\arguments{
\item{gblup_data}{GBLUPData object}
}
\value{
Invisible NULL
}
\description{
Prints comprehensive summary of a GBLUPData object.
}
\examples{
\dontrun{
gdata <- simulate_genotypes(n_ind = 100, n_snp = 500)
gdata <- simulate_phenotypes(gdata, n_env = 3, h2 = 0.6)
summary_gblup_data(gdata)
}
}
