% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_multi_trait.R
\name{summary_multi_trait}
\alias{summary_multi_trait}
\title{Summary Method for Multi-Trait Model}
\usage{
summary_multi_trait(mt_model)
}
\arguments{
\item{mt_model}{Multi-trait model from fit_multi_trait()}
}
\value{
Invisible NULL
}
\description{
Summary Method for Multi-Trait Model
}
