% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_gwas.R
\name{test_markers_vectorized}
\alias{test_markers_vectorized}
\title{Vectorized Marker Testing}
\usage{
test_markers_vectorized(
  y,
  X,
  M,
  K,
  Vi,
  sigma2_g,
  sigma2_e,
  method,
  P3D,
  n.cores,
  verbose
)
}
\description{
Vectorized Marker Testing
}
\keyword{internal}
