% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_multi_trait.R
\name{estimate_multivariate_varcomp}
\alias{estimate_multivariate_varcomp}
\title{Estimate Multivariate Variance Components}
\usage{
estimate_multivariate_varcomp(Y, K, nIters = 50, tol = 1e-04, verbose = FALSE)
}
\arguments{
\item{Y}{Matrix of trait values (individuals x traits)}

\item{K}{Relationship matrix}

\item{nIters}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
List with Vg and Ve matrices
}
\description{
Uses EM-REML to estimate genetic and residual covariance matrices
}
\keyword{internal}
