% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_multi_trait.R
\name{selection_index}
\alias{selection_index}
\title{Calculate Selection Index}
\usage{
selection_index(mt_model, weights)
}
\arguments{
\item{mt_model}{Multi-trait model from fit_multi_trait()}

\item{weights}{Named vector of weights for each trait}
}
\value{
Data frame with selection index values
}
\description{
Calculate Selection Index
}
